# letter_videos/openapi.json
{"openapi":"3.1.0","info":{"title":"AiPhonix Server","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/llm/chat":{"post":{"summary":"Llm Chat","operationId":"llm_chat_api_v1_llm_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/llm/logs":{"get":{"summary":"Llm Logs","operationId":"llm_logs_api_v1_llm_logs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/llm/logs/clear":{"post":{"summary":"Llm Logs Clear","operationId":"llm_logs_clear_api_v1_llm_logs_clear_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/llm/quiz":{"post":{"summary":"Llm Quiz","operationId":"llm_quiz_api_v1_llm_quiz_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/llm/quiz-generate":{"post":{"summary":"Llm Quiz Generate","description":"强制重新生成（忽略缓存）","operationId":"llm_quiz_generate_api_v1_llm_quiz_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizGenerateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tts/synthesize":{"post":{"summary":"Tts Synthesize","operationId":"tts_synthesize_api_v1_tts_synthesize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TTSRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/soe/evaluate":{"post":{"summary":"Soe Evaluate","operationId":"soe_evaluate_api_v1_soe_evaluate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SOERequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/wordbank/stats":{"get":{"summary":"Wordbank Stats","operationId":"wordbank_stats_api_v1_wordbank_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/wordbank/query":{"get":{"summary":"Wordbank Query","operationId":"wordbank_query_api_v1_wordbank_query_get","parameters":[{"name":"grade","in":"query","required":false,"schema":{"type":"string","default":"","title":"Grade"}},{"name":"semester","in":"query","required":false,"schema":{"type":"string","default":"","title":"Semester"}},{"name":"type","in":"query","required":false,"schema":{"type":"string","default":"","title":"Type"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/wordbank/search":{"get":{"summary":"Wordbank Search","operationId":"wordbank_search_api_v1_wordbank_search_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","default":"","title":"Q"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/wordbank/add-word":{"post":{"summary":"Wordbank Add Word","operationId":"wordbank_add_word_api_v1_wordbank_add_word_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddWordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/llm/word-info":{"post":{"summary":"Word Info","operationId":"word_info_api_v1_llm_word_info_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WordInfoRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/llm/sentence-generate":{"post":{"summary":"Sentence Generate","operationId":"sentence_generate_api_v1_llm_sentence_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SentenceGenerateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/llm/sentence-batch-save":{"post":{"summary":"Sentence Batch Save","operationId":"sentence_batch_save_api_v1_llm_sentence_batch_save_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SentenceBatchSaveRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chinese/polyphone":{"get":{"summary":"Chinese Polyphone","description":"多音字数据","operationId":"chinese_polyphone_api_v1_chinese_polyphone_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/practice/submit":{"post":{"summary":"Practice Submit","operationId":"practice_submit_api_v1_practice_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PracticeSession"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/practice/stats":{"get":{"summary":"Practice Stats","operationId":"practice_stats_api_v1_practice_stats_get","parameters":[{"name":"module","in":"query","required":false,"schema":{"type":"string","default":"","title":"Module"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/practice/history":{"get":{"summary":"Practice History","operationId":"practice_history_api_v1_practice_history_get","parameters":[{"name":"module","in":"query","required":false,"schema":{"type":"string","default":"","title":"Module"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/practice/char-record":{"post":{"summary":"Record Char Result","operationId":"record_char_result_api_v1_practice_char_record_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordCharResultRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/practice/char-weights":{"post":{"summary":"Get Char Weights","operationId":"get_char_weights_api_v1_practice_char_weights_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCharWeightsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/llm/word-suggestions":{"post":{"summary":"Word Suggestions","operationId":"word_suggestions_api_v1_llm_word_suggestions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WordSuggestionsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/english/vocabulary":{"get":{"summary":"English Vocabulary","operationId":"english_vocabulary_api_v1_english_vocabulary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/english/sentences":{"get":{"summary":"English Sentences","operationId":"english_sentences_api_v1_english_sentences_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/char-images":{"get":{"summary":"List Char Images","description":"查询汉字图片列表，支持按年级/学期/类型筛选","operationId":"list_char_images_api_v1_char_images_get","parameters":[{"name":"grade","in":"query","required":false,"schema":{"type":"string","default":"","title":"Grade"}},{"name":"semester","in":"query","required":false,"schema":{"type":"string","default":"","title":"Semester"}},{"name":"type_","in":"query","required":false,"schema":{"type":"string","default":"","title":"Type "}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":500,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/char-images/{char}":{"get":{"summary":"Get Char Image","description":"查询某个汉字的图片信息","operationId":"get_char_image_api_v1_char_images__char__get","parameters":[{"name":"char","in":"path","required":true,"schema":{"type":"string","title":"Char"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/char-images/file/{filename}":{"get":{"summary":"Serve Char Image","description":"提供图片文件","operationId":"serve_char_image_api_v1_char_images_file__filename__get","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AddWordRequest":{"properties":{"text":{"type":"string","title":"Text"},"type":{"type":"string","title":"Type","default":"chars"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]},"pinyin":{"type":"string","title":"Pinyin","default":""}},"type":"object","required":["text"],"title":"AddWordRequest"},"ChatRequest":{"properties":{"message":{"type":"string","title":"Message"},"mode":{"type":"string","title":"Mode","default":""},"prompt":{"type":"string","title":"Prompt","default":""}},"type":"object","required":["message"],"title":"ChatRequest"},"GetCharWeightsRequest":{"properties":{"chars":{"items":{"type":"string"},"type":"array","title":"Chars"}},"type":"object","required":["chars"],"title":"GetCharWeightsRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PracticeRecord":{"properties":{"char":{"type":"string","title":"Char","default":""},"module":{"type":"string","title":"Module","default":""},"attempt_count":{"type":"integer","title":"Attempt Count","default":0},"first_try_correct":{"type":"boolean","title":"First Try Correct","default":false},"hint_used":{"type":"boolean","title":"Hint Used","default":false},"correct":{"type":"boolean","title":"Correct","default":false},"timestamp":{"type":"string","title":"Timestamp","default":""},"date":{"type":"string","title":"Date","default":""},"grade":{"type":"string","title":"Grade","default":""}},"type":"object","title":"PracticeRecord"},"PracticeSession":{"properties":{"module":{"type":"string","title":"Module","default":""},"grade":{"type":"string","title":"Grade","default":""},"records":{"items":{"$ref":"#/components/schemas/PracticeRecord"},"type":"array","title":"Records","default":[]},"total_score":{"type":"integer","title":"Total Score","default":0},"max_score":{"type":"integer","title":"Max Score","default":0},"timestamp":{"type":"string","title":"Timestamp","default":""},"date":{"type":"string","title":"Date","default":""}},"type":"object","title":"PracticeSession"},"QuizGenerateRequest":{"properties":{"subtitle_text":{"type":"string","title":"Subtitle Text"},"video_name":{"type":"string","title":"Video Name"},"count":{"type":"integer","title":"Count","default":30}},"type":"object","required":["subtitle_text","video_name"],"title":"QuizGenerateRequest"},"QuizRequest":{"properties":{"subtitle_text":{"type":"string","title":"Subtitle Text"},"video_name":{"type":"string","title":"Video Name"},"count":{"type":"integer","title":"Count","default":30}},"type":"object","required":["subtitle_text","video_name"],"title":"QuizRequest"},"RecordCharResultRequest":{"properties":{"char":{"type":"string","title":"Char"},"correct":{"type":"boolean","title":"Correct","default":false},"type":{"type":"string","title":"Type","default":"pronunciation"}},"type":"object","required":["char"],"title":"RecordCharResultRequest"},"SOERequest":{"properties":{"ref_text":{"type":"string","title":"Ref Text"},"audio_base64":{"type":"string","title":"Audio Base64"},"engine":{"type":"string","title":"Engine","default":""}},"type":"object","required":["ref_text","audio_base64"],"title":"SOERequest"},"SentenceBatchSaveRequest":{"properties":{"sentences":{"additionalProperties":{"additionalProperties":true,"type":"object"},"type":"object","title":"Sentences"}},"type":"object","required":["sentences"],"title":"SentenceBatchSaveRequest"},"SentenceGenerateRequest":{"properties":{"word":{"type":"string","title":"Word"}},"type":"object","required":["word"],"title":"SentenceGenerateRequest"},"TTSRequest":{"properties":{"text":{"type":"string","title":"Text"},"speaker":{"type":"string","title":"Speaker","default":"0"},"speed":{"type":"integer","title":"Speed","default":5}},"type":"object","required":["text"],"title":"TTSRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WordInfoRequest":{"properties":{"char":{"type":"string","title":"Char"}},"type":"object","required":["char"],"title":"WordInfoRequest"},"WordSuggestionsRequest":{"properties":{"char":{"type":"string","title":"Char"}},"type":"object","required":["char"],"title":"WordSuggestionsRequest"}}}}